chore: fix other usages of single trace components
