+++ web/src/ee/features/annotation-queues/components/AnnotationQueueItemPage.tsx
@@ -116,45 +116,43 @@ const AnnotateIOView = ({
       }}
     >
       <ResizablePanel
-        className="col-span-1 h-full !overflow-y-auto"
+        className="col-span-1 h-full !overflow-y-auto rounded-md border"
         minSize={30}
         defaultSize={panelSize}
       >
         {view === "hideTree" ? (
-          item.objectType === AnnotationQueueObjectType.TRACE ? (
-            <TracePreview
-              key={trace.data.id}
-              trace={trace.data}
-              scores={trace.data.scores}
-              observations={trace.data.observations}
-              viewType="focused"
-              className="h-full"
-            />
-          ) : (
-            <ObservationPreview
-              observations={trace.data.observations}
-              scores={trace.data.scores}
-              projectId={item.projectId}
-              currentObservationId={item.objectId}
-              traceId={traceId}
-              viewType="focused"
-              className="h-full"
-            />
-          )
-        ) : (
-          <Card className="col-span-2 flex h-full flex-col overflow-hidden p-4">
-            <div className="overflow-x-auto md:overflow-hidden">
-              <Trace
+          <div className="max-h-full min-h-0 overflow-y-auto pl-4">
+            {item.objectType === AnnotationQueueObjectType.TRACE ? (
+              <TracePreview
                 key={trace.data.id}
                 trace={trace.data}
                 scores={trace.data.scores}
-                projectId={trace.data.projectId}
                 observations={trace.data.observations}
                 viewType="focused"
-                isValidObservationId={isValidObservationId}
               />
-            </div>
-          </Card>
+            ) : (
+              <ObservationPreview
+                observations={trace.data.observations}
+                scores={trace.data.scores}
+                projectId={item.projectId}
+                currentObservationId={item.objectId}
+                traceId={traceId}
+                viewType="focused"
+              />
+            )}
+          </div>
+        ) : (
+          <div className="max-h-full min-h-0 overflow-y-auto">
+            <Trace
+              key={trace.data.id}
+              trace={trace.data}
+              scores={trace.data.scores}
+              projectId={trace.data.projectId}
+              observations={trace.data.observations}
+              viewType="focused"
+              isValidObservationId={isValidObservationId}
+            />
+          </div>
         )}
       </ResizablePanel>
       <ResizableHandle withHandle className="ml-4 bg-transparent" />
