+++ web/src/features/trace-graph-view/components/TraceGraphView.tsx
@@ -1,8 +1,6 @@
 import React, { useEffect, useState, useMemo, useCallback } from "react";
 import { StringParam, useQueryParam } from "use-query-params";
-
 import type { APIScore, Trace } from "@langfuse/shared";
-import { Trace as TraceView } from "@/src/components/trace";
 import type { ObservationReturnTypeWithMetadata } from "@/src/server/api/routers/traces";
 
 import { TraceGraphCanvas } from "./TraceGraphCanvas";
