+++ web/src/styles/globals.css
@@ -54,7 +54,7 @@
     --dark-red: 0 72.2% 50.6%;
 
     --light-yellow: 55 91.7% 95.3%;
-    --dark-yellow: 32 81% 28.8%;
+    --dark-yellow: 43 96% 40%;
 
     --light-green: 141 84.2% 92.5%;
     --dark-green: 176 60.8% 19%;
@@ -114,9 +114,9 @@
     --primary-accent: 246 55% 70%;
     --muted-primary-accent: 239 83.5% 66.7%;
 
-    --muted-blue: 220, 70%, 25%;
-    --muted-green: 190, 70%, 25%;
-    --muted-magenta: 325, 70%, 25%;
+    --muted-blue: 215 95% 73.1%;
+    --muted-green: 142 76.7% 73.1%;
+    --muted-magenta: 330 76.7% 73.1%;
     --muted-gray: 210 50% 25%;
 
     --light-red: 0 62.8% 30.6%;
